requirements.txt
accelerate==0.25.0
actionlib==1.14.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
angles==1.9.13
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
dataclasses-json==0.6.2
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.9.0
docutils==0.20.1
dynamic-reconfigure==1.7.3
et-xmlfile==1.1.0
fastapi==0.104.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.0
fuzzywuzzy==0.18.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.4
idna==3.4
image-geometry==1.16.2
interactive-markers==1.12.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jellyfish==1.0.3
Jinja2==3.1.2
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonpatch==1.33
jsonpointer==2.4
jupyterlab-widgets==3.0.9
keyboard==0.13.5
kiwisolver==1.4.5
langchain==0.0.336
langsmith==0.0.64
laser_geometry==1.6.7
llvmlite==0.41.1
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
message-filters==1.16.0
more-itertools==10.1.0
moveit-commander==1.1.13
moveit-core==1.1.13
moveit-ros-planning-interface==1.1.13
moveit-ros-visualization==1.1.13
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.2.0
openai-whisper==20231117
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
py-cpuinfo==9.0.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydub==0.25.1
pyfiglet==1.0.2
pyparsing==3.1.1
python-dotenv==1.0.0
python-qt-binding==0.4.4
pytz==2023.3.post1
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
regex==2023.10.3
requests==2.31.0
resource_retriever==1.12.7
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.20
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
semantic-version==2.10.0
sensor-msgs==1.13.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools-rust==1.8.1
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sniffio==1.3.0
sounddevice==0.4.6
SQLAlchemy==2.0.23
srdfdom==0.6.4
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
topic-tools==1.16.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.225
urdfdom-py==0.4.6
urllib3==2.1.0
uvicorn==0.24.0.post1
Wave==0.0.2
Werkzeug==3.0.1
widgetsnbextension==4.0.9
xacro==1.14.17
yarl==1.9.2
